+++ src/main/java/general/Agent.java
@@ -5,6 +5,6 @@ public interface Agent {
   String getName();
   void initializeBeforeNewGame();
   Action chooseAction(State s);
-  void receiveReturn(int amount);
+  void receiveReturn(double amount);
   void gameOver();
 }
